requirements.txt
boto3==1.12.32
requests==2.24.0
python-dotenv==0.14.0

pytest==5.3.5
black==19.10b0
flake8==3.7.9
isort==5.1.0
coverage==5.0.4